pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hhw-brick"
version = "0.1.0"
description = "A comprehensive Python package for Heating Hot Water System (HHWS) Brick ontology applications"
readme = "README.md"
authors = [
    {name = "Mingchen Li", email = "liwei74123@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
keywords = ["brick", "building", "ontology", "hvac", "hot water", "rdf", "semantic web"]
requires-python = ">=3.8"
dependencies = [
    "rdflib>=6.0.0",
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
    "brickschema>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/CenterForTheBuiltEnvironment/HHW_brick"
Documentation = "https://github.com/CenterForTheBuiltEnvironment/HHW_brick/blob/main/README.md"
Repository = "https://github.com/CenterForTheBuiltEnvironment/HHW_brick"
Issues = "https://github.com/CenterForTheBuiltEnvironment/HHW_brick/issues"

[project.scripts]
hhw-brick = "hhw_brick.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["hhw_brick*"]
exclude = ["tests*", "docs*", "examples*", "scripts*"]

[tool.setuptools.package-data]
hhw_brick = [
    "*.ttl",
    "*.yaml",
    "*.yml",
    "conversion/*.yaml",
    "conversion/*.yml",
]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | htmlcov
  | __pycache__
)/
'''

# isort import sorting configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
skip = [".venv", "venv", ".tox", "build", "dist"]

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["hhw_brick"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Bandit security linter configuration
[tool.bandit]
exclude_dirs = ["tests", "examples", "scripts"]
skips = ["B101", "B601"]

# MyPy type checker configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "brickschema.*",
    "rdflib.*",
    "pandas.*",
    "numpy.*",
    "yaml.*",
]
ignore_missing_imports = true
